pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "modal-ssh"
version = "0.1.0"
description = "A simple python package for ssh'ing into a modal container"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "User", email = "user@example.com"},
]
dependencies = [
    "modal-client>=0.56.4",
]

[project.scripts]
modal-ssh = "modal_ssh.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["modal_ssh"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
] 